Register Controller fix

+++ src/utils/cloudinary.js
@@ -5,24 +5,30 @@ import fs from "fs"; //file system comes automatically from node
 cloudinary.config({ 
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
     api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env. CLOUDINARY_API_SECRET
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 const uploadonCloudinary = async (localFilePath)=>{
     try{
         
-        if(!localFilePath) return NULL;
+        if(!localFilePath)
+        {
+            console.log("No path found");
+            return NULL;
+        }
         //upload the file on cloudinary
-        const response = await cloudinary.v2.uploader.upload(localFilePath,{
+        const response = await cloudinary.uploader.upload(localFilePath,{
             resource_type:"auto",  
         });
         //file has been successfully uploaded
-        console.log("File is uploaded on cloudinary",response.url);
+        // console.log("File is uploaded on cloudinary",response.url);
+        fs.unlinkSync(localFilePath);
         console.log(response);
         return response;
 
     }catch(e){
-        fs.unlinkSync(localFilePath); //remove temporary file locally saved on our system as our operation got falied
+        console.log("Cloudinary per upload failed")
+        if(localFilePath) fs.unlinkSync(localFilePath); //remove temporary file locally saved on our system as our operation got falied
         return null;
     }
 }
